153 sometime the calculated lte throughput extracted from the network has negative values while it should be always = than txbytes (#154)
* chore: Update data types in LteEnbNetDevice

The commit updates the data types in the LteEnbNetDevice class to ensure consistency and avoid potential issues with type conversions.

* change data type to mmwave

* change times

* add metrics

* fix of the pdcp split

* save testing framework

* restart of the works

* test 7 cells

* Log NR RLC PDU and TX bytes through LTE
add indication periodicity as parameter to mmwave helper

* perform fix of periodicity also on PDCP

+++ src/mmwave/model/mmwave-enb-net-device.h
@@ -245,8 +245,8 @@ class MmWaveEnbNetDevice : public MmWaveNetDevice
     void RegisterNewSinrReading(uint64_t imsi, uint16_t cellId, long double sinr);
     std::map<uint64_t, std::map<uint16_t, long double>> m_l3sinrMap;
     uint64_t m_startTime;
-    std::map<uint64_t, uint32_t> m_drbThrDlPdcpBasedComputationUeid;
-    std::map<uint64_t, uint32_t> m_drbThrDlUeid;
+    std::map<uint64_t, double> m_drbThrDlPdcpBasedComputationUeid;
+    std::map<uint64_t, double> m_drbThrDlUeid;
     bool m_isReportingEnabled; //! true is KPM reporting cycle is active, false otherwise
     bool m_reducedPmValues;    //< if true use a reduced subset of pmvalues
